Refactor: extract base, normal and repeaterCard
extract base and normal to avoid circular import because repeaterCard has to import normal
