+++ src/widgets.js
@@ -1,187 +1,19 @@
 import collapse_toggle from "vellum/templates/collapse_toggle.html";
-import ui_element from "vellum/templates/ui_element.html";
 import widget_chips_template from "vellum/templates/widget_chips.html";
 import widget_control_keyvalue from "vellum/templates/widget_control_keyvalue.html";
-import widget_control_message from "vellum/templates/widget_control_message.html";
-import widget_repeater_card from "vellum/templates/widget_repeater_card.html";
-import nested_xpath_field from "vellum/templates/nested_xpath_field.html";
-import nested_dropdown_field from "vellum/templates/nested_dropdown_field.html";
 import _ from "underscore";
 import $ from "jquery";
 import atwho from "vellum/atwho";
 import util from "vellum/util";
 import richTextUtils from "vellum/richText";
-import nestedXPathField from "vellum/nestedXPathField";
 import analytics from "vellum/hqAnalytics";
-
-var base = function(mug, options) {
-    // set properties shared by all widgets
-    var widget = {};
-    options.richText = true;
-    widget.options = options;
-    widget.mug = mug;
-    widget.id = options.id;
-    util.eventuality(widget);
-
-    widget.isDisabled = function () {
-        // requires widget.path to be set.  This only happens in
-        // normal.  Need to change widgets that inherit directly from
-        // base to use the path/property system.
-        if (!widget.path) {
-            return false;
-        }
-
-        if (_.isFunction(mug.spec[widget.path].enabled)) {
-            return !mug.spec[widget.path].enabled(mug);
-        }
-
-        return mug.form.vellum.isPropertyLocked(mug, widget.path);
-    };
-
-    widget.getDisplayName = function () {
-        // use the display text, or the property name if none found
-        return this.definition.lstring ? this.definition.lstring : this.path;
-    };
-
-    widget.getControl = function () {
-        throw ("must be overridden");
-    };
-
-    widget.setValue = function (val) {
-        // noop
-    };
-
-    widget.getValue = function () {
-        // noop
-    };
-
-    widget.getHelp = function () {
-        if (this.definition && this.definition.help) {
-            return {
-                text: this.definition.help,
-                url: this.definition.helpURL
-            };
-        }
-        return null;
-    };
-
-    widget.refreshMessages = function () {
-        // placeholder to be overridden by widgets that inherit from base
-    };
-
-    widget.postRender = function () {
-        // placeholder to be overriden by widgets that inherit from base
-        // intended to allow post render configuration, such as
-        // enabling select2 functionality
-    };
-
-    widget.handleChange = function () {
-        widget.updateValue();
-        mug.showChangedMsg = true;
-        // TODO make all widgets that inherit from base set path
-        if (widget.path) {
-            // Widget change events, in addition to mug property
-            // setters, trigger mug validation because some mug
-            // property values have sub-properties that do not
-            // trigger mug property change events when they are
-            // changed.
-            mug.validate(widget.path);
-        }
-        widget.fire("change");
-    };
-
-    widget.updateValue = function () {
-        // When a widget's value changes, do whatever work you need to in
-        // the model/UI to make sure we are in a consistent state.
-        widget.save();
-    };
-
-    widget.save = function () {
-        throw 'Not Implemented';
-    };
-
-    widget.getUIElement = function () {
-        return getUIElement(widget.getControl(), widget.getDisplayName(),
-                            !!widget.isDisabled(), widget.getHelp());
-    };
-
-    widget.addInstanceRef = function (attrs, property) {
-        if (!property) {
-            property = widget.path;
-            if (!property) {
-                throw new Error("widget has no path: " + widget);
-            }
-        }
-        return mug.form.addInstanceIfNotExists(attrs, mug, property);
-    };
-
-    return widget;
-};
-
-var normal = function(mug, options) {
-    var path = options.widgetValuePath || options.path,
-        inputID = options.id || 'property-' + path,
-        disabled = options.disabled || false,
-        widget = base(mug, options);
-
-    widget.mugValue = options.mugValue || function (mug, value) {
-        if (arguments.length === 1) {
-            return mug.p[path];
-        }
-        mug.p[path] = value;
-    };
-
-    widget.path = path;
-    widget.definition = mug.p.getDefinition(options.path);
-    widget.currentValue = widget.mugValue(mug);
-    widget.id = inputID;
-    widget.saving = false;
-
-    widget.input = $("<input />")
-        .attr("name", inputID)
-        .attr("id", inputID)
-        .prop('disabled', disabled);
-
-    widget.getControl = function () {
-        return widget.input;
-    };
-
-    widget.getMessagesContainer = function () {
-        return widget.getControl()
-                .closest(".widget")
-                .find(".messages:last");
-    };
-
-    widget.getMessages = function (mug, path) {
-        return getMessages(mug, path);
-    };
-
-    widget.refreshMessages = function () {
-        var messages = widget.getMessages(mug, path);
-        var $container = widget.getMessagesContainer();
-        $container.empty();
-        if (messages.length) {
-            $container.append(messages);
-            $container.removeClass("hide");
-        } else {
-            $container.addClass("hide");
-        }
-    };
-
-    mug.on("messages-changed",
-           function () { widget.refreshMessages(); }, null, "teardown-mug-properties");
-
-    widget.save = function () {
-        widget.saving = true;
-        try {
-            widget.mugValue(mug, widget.getValue());
-        } finally {
-            widget.saving = false;
-        }
-    };
-
-    return widget;
-};
+import {
+    base,
+    normal,
+    getMessages,
+    getUIElement,
+} from "vellum/widgets/base";
+import repeaterCard from "vellum/widgets/repeaterCard";
 
 var text = function (mug, options) {
     var widget = normal(mug, options),
@@ -843,59 +675,6 @@ var getUIElementWithEditButton = function($uiElem, editFn, isDisabled) {
     return $uiElem;
 };
 
-var getUIElement = function($input, labelText, isDisabled, help) {
-    var $uiElem = $(ui_element({
-        labelText: labelText,
-        help: help,
-    }));
-
-    if (isDisabled) {
-        // Disable anything that can be disabled
-        $input.find("*").addBack().prop('disabled', true);
-        $input.filter('[contenteditable]').attr({
-            'contenteditable': false,
-            'disabled': true,
-        });
-    }
-
-    $uiElem.find(".controls").prepend($input);
-
-    if (help && !help.url) {
-        $uiElem.find(".fd-help a").click(function (e) { e.preventDefault(); });
-    }
-
-    return $uiElem;
-};
-
-function getMessages(mug, path) {
-    var $messages = $(),
-        seen = {};
-    mug.messages.each(path, function (msg) {
-        const messageKey = mug.messages.getMessageText(msg.message);
-        if (seen.hasOwnProperty(messageKey)) { return; }
-        seen[messageKey] = true;
-        let htmlMessage = "";
-        if (msg.message.hasOwnProperty("markdown")) {
-            htmlMessage = util.markdown(msg.message.markdown);
-        } else if (/n/.test(msg.message)) {
-            // html swallows newlines by default, so treat these messages as HTML to embed newline objects
-            htmlMessage = util.markdown(msg.message);
-        }
-
-        const context = {msg: msg, html: htmlMessage};
-        const html = $(widget_control_message(context));
-
-        html.find("button.close").click(function () {
-            mug.dropMessage(path, msg.key);
-            if (msg.key === "mug-nodeID-changed-warning") {
-                mug.showChangedMsg = false;
-            }
-        });
-        $messages = $messages.add(html);
-    });
-    return $messages;
-}
-
 function getWidget(input, vellum) {
     var obj = input,
         widget;
@@ -963,271 +742,6 @@ function enableAutocompleteOnInput($input, mug, options) {
         useRichText: mug.form.richText,
     });
 }
-
-// -------------------------------------------------------------------------
-// Repeater card widget — a compound-list property (as opposed to a scalar
-// one). Renders a list of cards (one per record); each record has N fields
-// declared by `cardConfig.fields`.
-
-// cardConfig:
-//   - fields: array of field specs (one row per field inside each card).
-//   - rootClass: class added to each `.fd-repeater-card`.
-//   - cardHeaderText: title shown in each card header.
-//   - addLabel: label for the "add row" action.
-//   - errorSummary (optional)
-//   - requiresAtLeastOne (optional)
-//   - emptyStateMessage (optional)
-//
-// cardConfig.fields:
-//   - label
-//   - fieldClass: control CSS class (also used by widgets.js selectors)
-//   - isIdentifier: when true, value comes from `cardIdentifier`
-//   - valueKey: when set, value comes from `cardData[valueKey]`
-//   - widget: "input" (default) | "xpath" | "dropdown"
-//   - required (optional): default to false
-//   - placeholder (optional)
-//   - options (optional): dropdown options (array or resolver)
-//   - extraValidator (optional)
-// -------------------------------------------------------------------------
-
-function readFieldValue($el) {
-    if (!$el.length) { return ""; }
-    var wrapper = $el.data("editorWrapper");
-    if (wrapper) { return wrapper.getValue(); }
-    return $el.val();
-}
-
-function validateField($field, mug, cardConfig) {
-    var $fieldRow = $field.closest('.form-group'),
-        $err = $fieldRow.find('.fd-field-error'),
-        required = $field.attr('data-required') === 'true',
-        widgetType = $field.attr('data-widget'),
-        val = readFieldValue($field),
-        touched = !!$field.data('touched'),
-        error = null;
-
-    if (required && !val && touched) {
-        error = gettext("Required");
-    } else if (widgetType === "xpath" && val) {
-        try {
-            mug.form.xpath.parse(val);
-        } catch (e) {
-            error = gettext("Invalid XPath expression");
-        }
-    }
-
-    if (!error && cardConfig) {
-        var fieldSpec = _.find(cardConfig.fields, function (f) {
-            return $field.hasClass(f.fieldClass);
-        });
-        if (fieldSpec && fieldSpec.extraValidator) {
-            error = fieldSpec.extraValidator(val);
-        }
-    }
-
-    if (error) {
-        $fieldRow.addClass('has-error');
-        $err.text(error).removeClass('hide');
-    } else {
-        $fieldRow.removeClass('has-error');
-        $err.text('').addClass('hide');
-    }
-}
-
-function emptyRepeaterItem(cardConfig) {
-    return _.reduce(cardConfig.fields, function (o, f) {
-        if (f.valueKey) { o[f.valueKey] = ""; }
-        return o;
-    }, {});
-}
-
-var repeaterCard = function (mug, options) {
-    var widget = normal(mug, options),
-        id = options.id || 'property-' + options.path,
-        cardConfig = options.cardConfig;
-    options.richText = false;
-
-    widget.input = $('<div class="control-row" />').attr('name', id);
-    widget.hasLogicReferences = true;
-
-    widget.getControl = function () {
-        return widget.input;
-    };
-
-    // Read values straight from the DOM using `cardConfig.fields` — no
-    // per-mug `getValue` override needed. Each card's key is the
-    // `isIdentifier` field's value; each card's body is
-    // {valueKey: fieldValue, ...}.
-    widget.getValue = function () {
-        var currentValues = {};
-        widget.input.find('.' + cardConfig.rootClass).each(function () {
-            var $card = $(this), key = null, entry = {};
-            _.each(cardConfig.fields, function (f) {
-                var val = readFieldValue($card.find('.' + f.fieldClass));
-                if (f.isIdentifier) { key = val; }
-                else if (f.valueKey) { entry[f.valueKey] = val; }
-            });
-            if (key !== null) { currentValues[key] = entry; }
-        });
-        return currentValues;
-    };
-
-    // Refresh inline field validation when logic.js updates references on
-    // external events (rename, delete). Without this the widget-level
-    // summary fires but the inline "Unknown question: X" text goes stale.
-    mug.on("messages-changed", function () {
-        widget.input.find('.fd-repeater-card').each(function () {
-            if (widget.validateCard) { widget.validateCard($(this)); }
-        });
-    }, null, "teardown-mug-properties");
-
-    // Empty value → render no cards (just the Add button). Clicking Add
-    // seeds a blank `""` entry via `addProperty`; widgets requiring at
-    // least one card surface a reminder via their `validationFunc`.
-    widget.refreshControl = function (value) {
-        value = value ? value : widget.getValue();
-        renderCards(value);
-        wirePlainInputHandlers();
-        wireXPathFields();
-        widget.input.find('.fd-add-property').click(widget.addProperty);
-        widget.input.find('.fd-remove-property').click(widget.removeProperty);
-        seedTouchedStateForSavedCards();
-        widget.input.find('.fd-repeater-card').each(function () {
-            widget.validateCard($(this));
-        });
-
-        function renderCards(val) {
-            var resolvedCardConfig = _.extend({}, cardConfig, {
-                fields: _.map(cardConfig.fields, function (f) {
-                    if (_.isFunction(f.options)) {
-                        return _.extend({}, f, {options: f.options(mug, options)});
-                    }
-                    return f;
-                }),
-            });
-            widget.input.html(widget_repeater_card({
-                props: val,
-                cardConfig: resolvedCardConfig,
-                useRichText: !!mug.form.richText,
-                nested_xpath_field: nested_xpath_field,
-                nested_dropdown_field: nested_dropdown_field,
-            }));
-        }
-
-        function wirePlainInputHandlers() {
-            widget.input.find('input, select').not('.fd-xpath-input')
-                .on('change keyup', function () {
-                    $(this).data('touched', true);
-                    widget.handleChange();
-                    widget.validateCard($(this).closest('.fd-repeater-card'));
-                });
-            widget.input.find('input[type="text"]').not('.fd-xpath-input')
-                .addClass('jstree-drop')
-                .each(function () { atwho.autocomplete($(this), mug); });
-        }
-
-        function wireXPathFields() {
-            widget.input.find('.fd-xpath-input').each(function () {
-                var $el = $(this),
-                    $group = $el.closest('.fd-nested-xpath-field');
-                nestedXPathField(mug, {
-                    $el: $el,
-                    $editButton: $group.find('.fd-xpath-edit'),
-                    initialValue: $el.attr('data-initial-value') || '',
-                    path: options.path,
-                    displayXPathEditor: options.displayXPathEditor,
-                }).on('change', function () {
-                    $el.data('touched', true);
-                    widget.handleChange();
-                    widget.validateCard($el.closest('.fd-repeater-card'));
-                });
-            });
-        }
-
-        // Cards loaded from saved data (key != "") are treated as "already
-        // seen" — all their fields start touched so existing broken data
-        // surfaces immediately.
-        function seedTouchedStateForSavedCards() {
-            widget.input.find('.fd-repeater-card').each(function () {
-                var $card = $(this),
-                    $name = $card.find('[data-is-identifier="true"]').first(),
-                    cardHasName = !!readFieldValue($name);
-                if (cardHasName) {
-                    $card.find('[data-widget]').each(function () {
-                        $(this).data('touched', true);
-                    });
-                }
-            });
-        }
-    };
-
-    widget.validateCard = function ($card) {
-        $card.find('[data-widget]').each(function () {
-            validateField($(this), mug, cardConfig);
-        });
-        widget.syncMugMessages();
-        widget.refreshMessages();
-    };
-
-    widget.syncMugMessages = function () {
-        var key = 'mug-' + options.path + '-error';
-        if (widget.input.find('.has-error').length > 0) {
-            mug.addMessage(options.path, {
-                key: key,
-                level: mug.ERROR,
-                message: cardConfig.errorSummary,
-            });
-        } else {
-            mug.dropMessage(options.path, key);
-        }
-    };
-
-    var $saveButton = options.vellum.data.core.saveButton.ui,
-        eventNamespace = '.fd-repeater-' + widget.id;
-    $saveButton.on('show.bs.popover' + eventNamespace, function () {
-        widget.input.find('[data-widget]').each(function () {
-            $(this).data('touched', true);
-        });
-        widget.input.find('.fd-repeater-card').each(function () {
-            widget.validateCard($(this));
-        });
-    });
-    mug.on('teardown-mug-properties', function () {
-        $saveButton.off(eventNamespace);
-    }, null, 'teardown-mug-properties');
-
-    widget.setValue = function (value) {
-        value = _.isUndefined(value) ? {} : value;
-        widget.refreshControl(value);
-    };
-
-    widget.updateValue = function () {
-        widget.save();
-    };
-
-    widget.removeProperty = function (e) {
-        e.preventDefault();
-        $(this).closest('.fd-repeater-card').remove();
-        widget.handleChange();
-        widget.syncMugMessages();
-    };
-
-    widget.addProperty = function (e) {
-        e.preventDefault();
-        var currentValues = widget.getValue();
-        // If there's already a blank card, focus it instead of adding another.
-        if (!("" in currentValues)) {
-            currentValues[""] = emptyRepeaterItem(cardConfig);
-            widget.refreshControl(currentValues);
-            widget.handleChange();
-        }
-        widget.input.find('.fd-repeater-card').last()
-            .find('input').first().focus();
-    };
-
-    return widget;
-};
-repeaterCard.trackLogicReferences = true;
 
 /**
  * Open the expression-editor modal with the standard arg bundle. Shared
